Stock Market Predictions w/ XGBoost in Under 25 Lines!
This is a brief tutorial on how to use the XGBoost package in Python to predict buy and sell signals for stocks in Python, in under 25 lines. There are around 175 lines of comments and imports, but only 25 lines of code. This is a tutorial meant for beginners, and has been kept very simple, Please check again for updates! 

XGBoost is a powerful library that is behind many Kaggle competition winners. Check them out at https://github.com/dmlc/xgboost.